dynamically use native or web apiFetch
